pyproject.toml
[project]
name = "job-recemneder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.0.0",
    "openai>=1.102.0",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
